requirements.txt
google-generativeai>=0.3.0
speechrecognition>=3.10.0
gtts>=2.3.1
python-dotenv>=1.0.0